change syntactic sugar && Need debug updateView in Example
