+++ Example/Example_iOS/Example_iOS.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:">
+      location = "self:Example_iOS.xcodeproj">
    </FileRef>
 </Workspace>
