+++ Example/Example_iOS/Example_iOS/Info.plist
