+++ Viscosity.xcodeproj/xcuserdata/WzxJiang.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -2,32 +2,24 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>SchemeUserState</key>
-	<dict>
-		<key>Viscosity.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>0</integer>
-		</dict>
-		<key>ViscosityTests.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>2</integer>
-		</dict>
-	</dict>
 	<key>SuppressBuildableAutocreation</key>
 	<dict>
-		<key>1E86C4951DF50D1E00F21E5E</key>
+		<key>OBJ_19</key>
 		<dict>
 			<key>primary</key>
 			<true/>
 		</dict>
-		<key>1EBEBC621E2C74D4009E3333</key>
+		<key>OBJ_20</key>
 		<dict>
 			<key>primary</key>
 			<true/>
 		</dict>
-		<key>1EC3BC381E0654FB00E66D1C</key>
+		<key>OBJ_28</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>OBJ_29</key>
 		<dict>
 			<key>primary</key>
 			<true/>
