+++ Viscosity.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:Example.xcodeproj">
+      location = "group:Example/Example_iOS/Example_iOS.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Viscosity.xcodeproj">
